feat(person): create person controller and refactory
