+++ src/pages/person/index.html
@@ -10,8 +10,8 @@
         </header>
         <main>
             <h1>Módulo para CRUD de Pessoa</h1>
-            <a href="person-create.html">Criar nova Pessoa</a> <br />
-            <a href="person-list.html">Listar Pessoas</a>
+            <a href="person-create/person-create.html">Criar nova Pessoa</a> <br />
+            <a href="person-list/person-list.html">Listar Pessoas</a>
         </main>
         <footer>
             <p>footer</p>
