+++ src/pages/person/person-create/person-create.html
@@ -1,8 +1,8 @@
 <html>
     <head>
         <title>Zuplae - Criar nova Pessoa</title>
-        <link rel="stylesheet" href="../../reset.css">
-        <link rel="stylesheet" href="../../style.css">
+        <link rel="stylesheet" href="../../../reset.css">
+        <link rel="stylesheet" href="../../../style.css">
     </head>
     <body>
         <header>
@@ -20,11 +20,11 @@
                 <button type="reset">Limpar</button>
             </form>
             <hr>
-            <a href="index.html">Voltar</a>
+            <a href="../index.html">Voltar</a>
         </main>
         <footer>
             <p>footer</p>
         </footer>
-        <script src="../../script-create.js"></script>
+        <script type="module" src="./person-create.js"></script>
     </body>
 </html>
