+++ src/pages/person/person-create/person-create.js
@@ -1,17 +1,4 @@
- document.querySelector('form').onsubmit = function(event) {
-    event.preventDefault(); //parar o evento de submit (impedir de recarregar)
-    let id = crypto.randomUUID();
-    let firstName = event.target.firstName.value;
-    let lastName = event.target.lastName.value;
-    let birthDate = event.target.birthDate.value;
+import { PersonController } from '../person-controller.js'
 
-    let person = { id, firstName, lastName, birthDate };
-    let personList = []
-    let listStorage = localStorage.getItem("personList")
-    if (listStorage) {
-        personList = JSON.parse(listStorage);
-    }
-    personList.push(person);
-    localStorage.setItem("personList", JSON.stringify(personList));
-    event.target.reset(); // Limpar o formulário
-};
+let controller = new PersonController();
+controller.bindForm();
