+++ src/pages/person/person-list/person-list.html
@@ -1,8 +1,8 @@
 <html>
     <head>
         <title>Zuplae - Listagem de Pessoas</title>
-        <link rel="stylesheet" href="../../reset.css">
-        <link rel="stylesheet" href="../../style.css">
+        <link rel="stylesheet" href="../../../reset.css">
+        <link rel="stylesheet" href="../../../style.css">
     </head>
     <body>
         <header>
@@ -27,11 +27,11 @@
                 </tbody>                
             </table>
             <hr>
-            <a href="index.html">Voltar</a>
+            <a href="../index.html">Voltar</a>
         </main>
         <footer>
             <p>footer</p>
         </footer>
-        <script src="../../script-list.js"></script>
+        <script type="module" src="./person-list.js"></script>
     </body>
 </html>
